+++ moneyucab-portalweb-back/EntitiesForm/BilleteraTarjeta.cs
@@ -11,7 +11,7 @@ namespace moneyucab_portalweb_back.EntitiesForm
 		public int idUsuario { get; set; }
 		public int idTipoTarjeta { get; set; }
 		public int idBanco { get; set; }
-		public int numero { get; set; }
+		public long numero { get; set; }
 		public int ano { get; set; }
 		public int mes { get; set; } 
 		public int dia { get; set; }
